Entity 생성 후 refactoring 실시 (#27)
* ci : gitignore 수정

* refactor : 각 Entity별 상속관계 정리 및 setter 삭제
