+++ .gitignore
@@ -36,4 +36,5 @@ out/
 ### VS Code ###
 .vscode/
 
-.application-secret.yml
+application-secret.yml
+db_dev.mv.db
