+++ src/main/java/com/example/be8arm/domain/chat/chatRoom/entity/ChatRoom.java
@@ -1,6 +1,5 @@
-package com.example.be8arm.chat.chatRoom.entity;
+package com.example.be8arm.domain.chat.chatRoom.entity;
 
-import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
 
 import java.time.LocalDateTime;
@@ -11,39 +10,33 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import com.example.be8arm.chat.chatMessage.entity.ChatMessage;
+import com.example.be8arm.domain.chat.chatMessage.entity.ChatMessage;
+import com.example.be8arm.global.IdEntity;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.OrderBy;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
+import lombok.experimental.SuperBuilder;
 
 @Entity
 @Setter
 @Getter
 @AllArgsConstructor(access = PROTECTED)
 @NoArgsConstructor(access = PROTECTED)
-@Builder
+@SuperBuilder
 @ToString(callSuper = true)
 @EntityListeners(AuditingEntityListener.class)
-public class ChatRoom {
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-	@EqualsAndHashCode.Include
-	private Long id;
-
+public class ChatRoom extends IdEntity {
 	@CreatedDate
 	@Getter
 	private LocalDateTime createDate;
