+++ src/main/java/com/example/be8arm/domain/member/member/entity/Member.java
@@ -1,32 +1,28 @@
 package com.example.be8arm.domain.member.member.entity;
 
+import com.example.be8arm.global.TimeEntity;
+
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.OneToOne;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.experimental.SuperBuilder;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Builder
-@AllArgsConstructor
-public class Member {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long memberId;
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@SuperBuilder
+public class Member extends TimeEntity {
 
 	private String loginId;  // 로그인 ID
+
 	private String password;
+
 	private String name; // 실명
+
 	private String nickname;
 
 	private String imgUrl; // 프로필 사진
