+++ src/main/java/com/example/be8arm/domain/member/member/entity/Profile.java
@@ -1,13 +1,15 @@
 package com.example.be8arm.domain.member.member.entity;
 
-import com.example.be8arm.domain.global.TimeEntity;
+import com.example.be8arm.global.TimeEntity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.OneToOne;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+@Getter
 @SuperBuilder
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,10 +17,16 @@ import lombok.experimental.SuperBuilder;
 public class Profile extends TimeEntity {
 	@OneToOne
 	private Member member;
+
 	private Integer age;
+
 	private Gender gender;
+
 	private Integer benchPress;
+
 	private Integer deadLift;
+
 	private Integer squat;
+
 	private Integer totalWeight;
 }
